Add check for missing plugin when deleting, just logs at debug for now if not exist PLUG-469
git-svn-id: https://studio.atlassian.com/svn/PLUG/trunk@47203 2c54a935-e501-0410-bc05-97a93f6bca70

+++ atlassian-plugins-core/src/main/java/com/atlassian/plugin/loaders/DirectoryScanner.java
@@ -171,8 +171,17 @@ class DirectoryScanner implements com.atlassian.plugin.loaders.classloading.Scan
 
     public void remove(DeploymentUnit unit) throws PluginException
     {
-        if (!unit.getPath().delete())
-            throw new PluginException("Unable to delete file: " + unit.getPath());
+        if (unit.getPath().exists())
+        {
+            if (!unit.getPath().delete())
+            {
+                throw new PluginException("Unable to delete file: " + unit.getPath());
+            }
+        }
+        else
+        {
+            log.debug("Plugin file " + unit.getPath().getPath() + " doesn't exist to delete.  Ignoring.");
+        }
 
         clear(unit.getPath());
     }
